pyproject.toml
[project]
name = "privacy-disclosure-gap"
version = "0.1.0"
description = "국내 스타트업 웹사이트의 개인정보 처리방침 컴플라이언스 및 Disclosure-Practice Gap 분석 시스템"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.42.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "pandas>=2.2.1",
    "pydantic>=2.6.4",
    "typer>=0.12.3",
    "scipy>=1.12.0",
    "statsmodels>=0.14.1",
    "matplotlib>=3.8.3",
    "pyyaml>=6.0.1",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
privacy-gap = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
